Remove overloading of function mpi_get_size
The function mpi_get_size in mpi_utils.cc was renamed to include the
name of the type it is operating on, i.e.: mpi_get_comm_size,
mpi_get_group_size and mpi_get_type_size. Overloading does not work on
MPI implementations where the MPI_Comm and MPI_Group types are the
same.